+++ src/main/java/io/supertokens/pluginInterface/totp/exception/TotpNotEnabledException.java
@@ -1,0 +1,5 @@
+package io.supertokens.pluginInterface.totp.exception;
+
+public class TotpNotEnabledException extends Exception {
+    private static final long serialVersionUID = 6848053563771647272L;
+}